use inner static class
